Remove VCS fallback functionality
