+++ go.mod
@@ -5,7 +5,6 @@ go 1.15
 require (
 	github.com/jfrog/gofrog v1.0.6
 	github.com/jfrog/jfrog-client-go v0.25.0
-	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.7.0
 )
